prayfile: handle cp1252/utf8 encoding correctly
The original Creatures engines use the CP-1252 encoding (you can
see this in e.g. WhiteOnTransparentChars.s16, which maps over
perfectly).

Modern practices, however, are to store files and operate on
strings in UTF-8. So, in this commit we do a couple things:

1) When parsing Pray or CAOS2Pray sources, convert them to UTF-8
   if they're not already
2) When reading Pray files, convert strings to UTF-8, if they're
   not already
3) When writing Pray files, convert strings to CP-1252, if they're
   not already

This follows the Robustness Principle of "be conservative in what
you do, be liberal in what you accept."