SFCFile: fix bug where signed integers are converted to unsigned
Exhibited as an agent in C2 with an x velocity of ~400000000 (which
should've been a small negative number).

Bug appeared after 2ff6bb16 which changed Agent::velx from a caosValue
to a float- previously the crazily large unsigned x velocity would be
implicitly converted to a signed int to fit inside a caosValue.

Consider turning on -Wsign-conversion.