HIST EVNT: don't error when called when importing starter pair of norns
In C3 and DS, when choosing to start with a couple of pre-trained adult
norns, HIST EVNT va00 0 "" "" is called even though the moniker doesn't
exist. DS claims it's to "doctor history so reconciliation fails".

Importing norns doesn't work anyways, but make the error disappear.

TODO: what is this actually supposed to do? delete the history maybe?