endianlove: use byteswapping because it's much faster
This file used to handle integer encodings in a platform-neutral
way using byte shifts and ors, in the same vein as Rob Pike's
article "The Byte Order Fallacy":

e.g. a little-endian uint32 would be read as (buf[0] << 0) | (buf[1] << 8) | (buf[2] << 16) | (buf[3] << 24);

Unfortunately, this is not optimized out on all platforms
(*cough*MSVC*cough*), and even when it is optimized for the
singular case it does not optimize well for the "readmany" case.

So, we do platform endianness detection and byte swapping.