C1Sound: fix volume being overridden by C1SoundManager when moving listener
Problem exposed in 32c18dbf which set the background music volume to 50%.

When scrolling the window, the background music would start "warbling" and
have distortion / crackling.

What was happening was C1MusicManager would set the volume to 50% every frame,
but C1SoundManager would set it back to 100% every tick (~every 6 frames)
when the viewport changed.