Fix out of bound read in CaosLexer
I noticed that the CaosLexer has some bugs where it reads beyond the null termination character for malformed input. This can cause a read into unallocated memory, which can potentially cause the code to crash.

There's a security concern here, as malicious input can cause the lexer to read uninitialized memory (similar to [Heartbleed](https://en.wikipedia.org/wiki/Heartbleed)), though it's only a single byte, and it's hard to imagine an attacker weaponizing this in the context of openc2e.

This change does two things:

1. Adds a unit test to exercise the previously vulnerable code path. You can't see the failure since the fix is also in this PR, but you can see what happens without the fix [in this build on my copy of the repo](https://github.com/mtlynch/openc2e/actions/runs/11849613968/job/33023157739)
2. Fixes the vulnerability by gracefully handling the null terminator following a double quote character