Print clang-format version during CI build
For tools that are highly-sensitive to exact versions, I find it helpful to have the exact version in the build logs to know how to reproduce the CI steps locally.

I realize there's a 10 in the binary name, but I actually didn't realize that was the version. But beyond that, I think having the exact version is helpful and has no real downside.