In CAOS lexer, increment line number for newlines within byte string
Previously, the if the lexer encountered a newline or carriage return within a bytestring, it wouldn't increment the line count of the tokens even though they technically appeared on a later line of input. This updates the line counter so that newlines or carriage returns within a byte string increment the line count of the lexer's tokens.